remove dublicate code and client side changed apply to backend
