+++ client/src/Pages/Settings/index.jsx
@@ -93,12 +93,9 @@ const Settings = () => {
 
 		if (name === "timezone") {
 			dispatch(setTimezone({ timezone: value }));
-			// Also save to database settings
-			const newSettingsData = {
-				...settingsData,
-				settings: { ...settingsData.settings, timezone: value },
-			};
 			setSettingsData(newSettingsData);
+			// Save timezone change to server immediately
+			saveSettings(newSettingsData.settings);
 			return;
 		}
